+++ lib/asc_tooling/pagination.rb
@@ -1,13 +1,30 @@
 module ASCTooling
   module Pagination
     def paginated_resources(path, params: {}, limit: Client::DEFAULT_PAGE_LIMIT)
-      resources = []
+      paginated_document(path, params: params, limit: limit).fetch("data")
+    end
+
+    def paginated_document(path, params: {}, limit: Client::DEFAULT_PAGE_LIMIT)
+      document = { "data" => [], "included" => [] }
+
+      each_paginated_response(path, params: params, limit: limit) do |response|
+        document["data"].concat(response.fetch("data", []))
+        document["included"].concat(response.fetch("included", []))
+      end
+
+      document["included"].uniq! { |resource| [resource["type"], resource["id"]] }
+      document
+    end
+
+    private
+
+    def each_paginated_response(path, params:, limit:)
       next_path = path
       next_params = params.merge("limit" => limit.to_s)
 
       loop do
         response = request_json("GET", next_path, params: next_params)
-        resources.concat(response.fetch("data", []))
+        yield response
 
         next_link = response.dig("links", "next")
         break if Client.blank?(next_link)
@@ -16,8 +33,6 @@ module ASCTooling
         next_path = uri.path
         next_params = URI.decode_www_form(uri.query || "").to_h
       end
-
-      resources
     end
   end
 end